ints

+++ Lasertag Modular/Assets/Scripts/Network/Packets/Protocol.cs
@@ -58,32 +58,32 @@
 
     public class SetupMobile
     {
-        public byte gameId;
+        public int gameId;
         public byte playerId;
         public Characters character;
     }
 
     public class SetupVest
     {
-        public byte gameId;
+        public int gameId;
         public byte playerId;
     }
 
     public class SetupWeapon
     {
-        public byte gameId;
+        public int gameId;
         public byte playerId;
     }
 
     public class SetupGrenade
     {
-        public byte gameId;
+        public int gameId;
         public byte playerId;
     }
 
     public class SetupCar
     {
-        public byte gameId;
+        public int gameId;
         public byte playerId;
         public string ipAddress;
         public int portToListen;
